+++ scripts/tables.h
@@ -35,15 +35,15 @@ extern void printCSVDataTableInt(const char *filename, DataTableInt *table, bool
 typedef struct distance_matrix_s {
 	size_t numObs;
 	
-	double *distances;
+	long long unsigned int *distances;
 	char **obsNames;
 } DistanceMatrix;
 
 extern DistanceMatrix *allocDistanceMatrix(size_t numVars);
 
-extern void setDistanceMatrix(DistanceMatrix *mat, size_t obsNum1, size_t obsNum2, double val);
+extern void setDistanceMatrix(DistanceMatrix *mat, size_t obsNum1, size_t obsNum2, long long unsigned int val);
 
-extern double getDistanceMatrix(DistanceMatrix *mat, size_t obsNum1, size_t obsNum2);
+extern long long unsigned int getDistanceMatrix(DistanceMatrix *mat, size_t obsNum1, size_t obsNum2);
 
 extern void setDistanceMatrixObsName(DistanceMatrix *mat, size_t obsNum, const char *name);
