+++ scripts/tau.c
@@ -8,7 +8,7 @@
 //Computes the Kendall's tau Hausdorff distance between partial rankings `a` and `b`.
 //Each partial ranking is of `length` items.
 //Not every ranking need be present in `a` or `b`.
-double tauHausdorffDistance(const int *a, const int *b, size_t length, size_t numRanks) {
+long long unsigned int tauHausdorffDistance(const int *a, const int *b, size_t length, size_t numRanks) {
 	NijMatrix *nij = getNij(a, b, length, numRanks);
 	
 	freeNijMatrix(nij);
